Created an abstract Answer class and refactored project using MVC
