+++ src/main/java/JpaApplication/Controller/AnswerController.java
@@ -1,5 +1,7 @@
-package JpaApplication;
+package JpaApplication.Controller;
 
+import JpaApplication.Model.Answer;
+import JpaApplication.Repository.AnswerRepository;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
