+++ src/main/java/JpaApplication/Controller/QuestionController.java
@@ -1,11 +1,8 @@
-package JpaApplication;
+package JpaApplication.Controller;
 
-import org.springframework.web.bind.annotation.DeleteMapping;
+import JpaApplication.Model.Question;
+import JpaApplication.Repository.QuestionRepository;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
