+++ src/main/java/JpaApplication/Controller/SurveyController.java
@@ -1,5 +1,11 @@
-package JpaApplication;
+package JpaApplication.Controller;
 
+import JpaApplication.Model.Answer;
+import JpaApplication.Model.OpenEnded;
+import JpaApplication.Model.Question;
+import JpaApplication.Model.Survey;
+import JpaApplication.Repository.QuestionRepository;
+import JpaApplication.Repository.SurveyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -26,10 +32,8 @@ public class SurveyController {
     @PostMapping("/addQuestion")
     public String addQuestionSubmit(@ModelAttribute Question question, Model model) {
         model.addAttribute("question", question);
-        Answer answer = new Answer("answer");
-        question.setAnswersByType("Open-Ended");
+        Answer answer = new OpenEnded("answer");
         question.addAnswer(answer);
-//
         QuestionRepo.save(question);
 
         Question q1 = QuestionRepo.findByQuestions(question.getQuestions());
