+++ src/main/java/JpaApplication/Model/AllUsers.java
@@ -1,4 +1,4 @@
-package JpaApplication;
+package JpaApplication.Model;
 
 import javax.persistence.*;
