+++ src/main/java/JpaApplication/Repository/AnswerRepository.java
@@ -1,5 +1,6 @@
-package JpaApplication;
+package JpaApplication.Repository;
 
+import JpaApplication.Model.Answer;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
