+++ src/main/java/JpaApplication/Repository/SurveyRepository.java
@@ -1,5 +1,6 @@
-package JpaApplication;
+package JpaApplication.Repository;
 
+import JpaApplication.Model.Survey;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
